pyproject.toml
[project]
name = "deepscribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.3.0",
    "pydantic>=2.12.3",
    "pydantic-ai>=1.8.0",
    "nltk>=3.8.1",
    "torch>=2.3.0",
    "transformers>=4.40.0",
    "tiktoken>=0.12.0",
    "pydantic-evals>=1.8.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.0",
]
